pm-agent/requirements.txt
anthropic==0.30.0
google-genai>=1.0.0
python-dotenv==1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
PyJWT>=2.8.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sentry-sdk[fastapi]>=1.40.0
prometheus-client>=0.20.0
pytest-asyncio>=0.23.0
